backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.*
uvicorn[standard]==0.34.*
gunicorn==23.*          # Production WSGI/ASGI server
python-multipart==0.0.* # Form / file upload support

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*         # Async PostgreSQL driver
psycopg2-binary==2.9.*  # Sync driver (required by Alembic)
alembic==1.14.*         # Database migrations

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic==2.10.*
pydantic-settings==2.7.*

# ── Cache / Task Queue ────────────────────────────────────────────────────────
celery[redis]==5.4.*
redis==5.2.*

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.28.*

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
email-validator==2.2.*

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.*
